+++ packages/@enotion/components/src/index.ts
@@ -1,3 +1,4 @@
 import "./index.css";
 
 export * from "./Skeleton.js";
+export * from "./Button.js";